requirements.txt
apache-airflow-providers-http==4.13.1
apache-airflow-providers-google==12.0.0
google-cloud-storage==2.19.0
